+++ src/classes/monitor.ts
@@ -21,9 +21,25 @@ export default class Monitor {
     return message.data.map((slot) => {
       switch (slot.type) {
         case 'player_id': {
-          const playerName = this.client.players.get(parseInt(slot.text))?.name
+          const playerId = parseInt(slot.text)
+          const playerName = this.client.players.get(playerId)?.name
           if (playerName && linkMap.has(playerName)) {
-            return `<@${linkMap.get(playerName)}>`
+            let shouldMention = true
+            if (message.type === 'ItemSend') {
+              if (playerId === (message as any).receiving) {
+                shouldMention = this.data.mention_item_receiver
+              } else {
+                shouldMention = this.data.mention_item_finder
+              }
+            } else if (message.type === 'Hint') {
+              shouldMention = this.data.mention_hints
+            } else if (message.type === 'Collect') {
+              shouldMention = this.data.mention_item_finder
+            }
+
+            if (shouldMention) {
+              return `<@${linkMap.get(playerName)}>`
+            }
           }
           return `**${playerName}**`
         }
@@ -143,9 +159,9 @@ export default class Monitor {
     const links = await Database.getLinks(this.guild.id)
     const linkMap = new Map<string, string>(links.map(l => [l.archipelago_name, l.discord_id]))
 
-    const formatPlayer = (slot: number) => {
+    const formatPlayer = (slot: number, mentionFlag: boolean = true) => {
       const playerName = this.client.players.get(slot)?.name
-      if (playerName && linkMap.has(playerName)) {
+      if (playerName && linkMap.has(playerName) && mentionFlag) {
         return `<@${linkMap.get(playerName)}>`
       }
       return `**${playerName}**`
@@ -163,14 +179,20 @@ export default class Monitor {
         // Overrides for special join messages
         if (packet.tags.includes('Monitor')) return
         if (packet.tags.includes('IgnoreGame')) {
-          this.send(`A tracker for ${formatPlayer(packet.slot)} has joined the game!`)
+          this.send(`A tracker for ${formatPlayer(packet.slot, this.data.mention_join_leave)} has joined the game!`)
           return
         }
 
-        this.send(`${formatPlayer(packet.slot)} (${this.client.players.get(packet.slot)?.game}) joined the game!`)
+        this.send(`${formatPlayer(packet.slot, this.data.mention_join_leave)} (${this.client.players.get(packet.slot)?.game}) joined the game!`)
         break
       case 'Part':
-        this.send(`${formatPlayer(packet.slot)} (${this.client.players.get(packet.slot)?.game}) left the game!`)
+        this.send(`${formatPlayer(packet.slot, this.data.mention_join_leave)} (${this.client.players.get(packet.slot)?.game}) left the game!`)
+        break
+      case 'Goal':
+        this.send(`${formatPlayer(packet.slot, this.data.mention_completion)} has completed their goal!`)
+        break
+      case 'Release':
+        this.send(`${formatPlayer(packet.slot, this.data.mention_item_finder)} has released their remaining items!`)
         break
     }
   }