+++ src/classes/monitordata.ts
@@ -4,4 +4,9 @@ export default class MonitorData {
   game: string
   player: string
   channel: string
+  mention_join_leave: boolean
+  mention_item_finder: boolean
+  mention_item_receiver: boolean
+  mention_completion: boolean
+  mention_hints: boolean
 }